+++ tgbotapi.extensions.api/src/commonMain/kotlin/dev/inmo/tgbotapi/extensions/api/bot/DeleteMyCommands.kt
@@ -1,11 +1,11 @@
 package dev.inmo.tgbotapi.extensions.api.bot
 
 import dev.inmo.tgbotapi.bot.TelegramBot
-import dev.inmo.tgbotapi.requests.bot.GetMyCommands
+import dev.inmo.tgbotapi.requests.bot.DeleteMyCommands
 import dev.inmo.tgbotapi.types.commands.BotCommandScope
 import dev.inmo.tgbotapi.types.commands.BotCommandScopeDefault
 
-suspend fun TelegramBot.getMyCommands(
+suspend fun TelegramBot.deleteMyCommands(
     scope: BotCommandScope = BotCommandScopeDefault,
     languageCode: String? = null
-) = execute(GetMyCommands(scope, languageCode))
+) = execute(DeleteMyCommands(scope, languageCode))